requirements.txt
flask
waitress
sentence-transformers
faiss-cpu
numpy
pypdf
pymupdf
requests
beautifulsoup4
